fix: rename file names & transaction amount
